Accessibility changes in file-input (#1246)
