+++ it/content/src/main/content/jcr_root/content/dam/formsanddocuments/core-components-it/samples/fileinput/fileinputv1/basic/.content.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<jcr:root xmlns:jcr="http://www.jcp.org/jcr/1.0" xmlns:dam="http://www.day.com/dam/1.0" xmlns:nt="http://www.jcp.org/jcr/nt/1.0" xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:fd="http://www.adobe.com/aemfd/fd/1.0"
+          jcr:primaryType="dam:Asset">
+    <jcr:content
+            jcr:lastModified="{Date}2023-01-25T17:24:29.344+05:30"
+            jcr:primaryType="dam:AssetContent"
+            sling:resourceType="fd/fm/af/render"
+            guide="1"
+            type="guide">
+        <metadata
+                fd:version="2.1"
+                jcr:primaryType="nt:unstructured"
+                allowedRenderFormat="HTML"
+                author="admin"
+                availableInMobileApp="{Boolean}false"
+                dorTemplateChanged="Boolean"
+                dorType="none"
+                formmodel="none"
+                hasCustomThumbnail="{Boolean}false"
+                themeRef="/libs/fd/af/themes/canvas"
+                title="Adaptive Form V1 (IT)"/>
+    </jcr:content>
+</jcr:root>
