+++ it/content/src/main/content/jcr_root/content/forms/af/core-components-it/samples/fileinput/fileinputv2/basic/.content.xml
@@ -1,0 +1,109 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<jcr:root xmlns:jcr="http://www.jcp.org/jcr/1.0" xmlns:nt="http://www.jcp.org/jcr/nt/1.0" xmlns:cq="http://www.day.com/jcr/cq/1.0" xmlns:sling="http://sling.apache.org/jcr/sling/1.0" xmlns:fd="http://www.adobe.com/aemfd/fd/1.0"
+          jcr:primaryType="cq:Page">
+    <jcr:content
+            cq:deviceGroups="[mobile/groups/responsive]"
+            cq:lastModified="{Date}2023-01-25T17:24:29.344+05:30"
+            cq:lastModifiedBy="admin"
+            cq:template="/conf/core-components-examples/settings/wcm/templates/af-blank-v2"
+            jcr:language="en"
+            jcr:primaryType="cq:PageContent"
+            jcr:title="Adaptive Form V2 (IT)"
+            sling:configRef="/conf/forms/core-components-it/samples/fileinput/fileinputv2/basic/"
+            sling:resourceType="forms-core-components-it/components/page-fileinputv2">
+        <guideContainer
+                fd:version="2.1"
+                jcr:lastModified="{Date}2023-01-17T16:29:57.778+05:30"
+                jcr:lastModifiedBy="admin"
+                jcr:primaryType="nt:unstructured"
+                sling:resourceType="forms-components-examples/components/form/container"
+                actionType="forms-core-components-it/customsubmission/logsubmit"
+                dorType="none"
+                fieldType="form"
+                prefillService="Core Custom Pre-fill Service"
+                textIsRich="true"
+                thankYouMessage="Thank you for submitting the form."
+                thankYouOption="page">
+            <fileinput1
+                    jcr:lastModified="{Date}2023-01-24T16:47:36.882+05:30"
+                    jcr:lastModifiedBy="admin"
+                    jcr:primaryType="nt:unstructured"
+                    jcr:title="File Input - 1"
+                    sling:resourceType="core/fd/components/form/fileinput/v2/fileinput"
+                    accept="[audio/*, video/*, image/*, text/*, application/pdf]"
+                    buttonText="Attach Files"
+                    enabled="{Boolean}true"
+                    fieldType="file-input"
+                    multiSelection="true"
+                    name="fileinput1"
+                    readOnly="{Boolean}false"
+                    textIsRich="[true,true]"
+                    type="file[]"
+                    visible="{Boolean}true"/>
+            <fileinput2
+                    jcr:primaryType="nt:unstructured"
+                    jcr:title="File Input - 2"
+                    sling:resourceType="core/fd/components/form/fileinput/v2/fileinput"
+                    accept="[audio/*, video/*, image/*, text/*, application/pdf]"
+                    fieldType="file-input"
+                    name="fileinput2"
+                    type="file"/>
+            <fileinput3
+                    jcr:primaryType="nt:unstructured"
+                    jcr:title="File Input - 3"
+                    sling:resourceType="core/fd/components/form/fileinput/v2/fileinput"
+                    accept="[application/pdf]"
+                    description="This is long description"
+                    fieldType="file-input"
+                    id="fileinput_tooltip_scenario_test"
+                    name="fileinput3"
+                    tooltip="This is short description"
+                    tooltipVisible="true"
+                    type="file"/>
+            <fileinput4
+                    jcr:primaryType="nt:unstructured"
+                    jcr:title="File Input - 4"
+                    sling:resourceType="core/fd/components/form/fileinput/v2/fileinput"
+                    accept="[application/x-msdownload,application/vnd.ms-outlook,application/pdf]"
+                    description="This is long description"
+                    fieldType="file-input"
+                    id="fileinput_unknown_mimetype_test"
+                    name="fileinput4"
+                    tooltip="This is short description"
+                    multiSelection="true"
+                    type="file"/>
+             <fileinput5
+                    jcr:lastModified="{Date}2023-01-24T16:47:36.882+05:30"
+                    jcr:lastModifiedBy="admin"
+                    jcr:primaryType="nt:unstructured"
+                    jcr:title="File Input - 5"
+                    sling:resourceType="core/fd/components/form/fileinput/v2/fileinput"
+                    accept="[audio/*, video/*, image/*, text/*, application/pdf]"
+                    buttonText="Attach Files"
+                    enabled="{Boolean}true"
+                    fieldType="file-input"
+                    name="fileinput5"
+                    readOnly="{Boolean}true"
+                    textIsRich="[true,true]"
+                    type="file[]"
+                    visible="{Boolean}true"/>
+            <submit
+                    jcr:lastModified="{Date}2023-01-17T16:28:58.844+05:30"
+                    jcr:lastModifiedBy="admin"
+                    jcr:primaryType="nt:unstructured"
+                    jcr:title="Submit"
+                    sling:resourceType="forms-components-examples/components/form/actions/submit"
+                    buttonType="submit"
+                    dorExclusion="true"
+                    fieldType="button"
+                    name="submit1673953138924">
+                <fd:rules
+                        fd:click="[{&quot;nodeName&quot;:&quot;ROOT&quot;\,&quot;items&quot;:[{&quot;nodeName&quot;:&quot;STATEMENT&quot;\,&quot;choice&quot;:{&quot;nodeName&quot;:&quot;EVENT_SCRIPTS&quot;\,&quot;items&quot;:[{&quot;nodeName&quot;:&quot;EVENT_CONDITION&quot;\,&quot;choice&quot;:{&quot;nodeName&quot;:&quot;EVENT_AND_COMPARISON&quot;\,&quot;items&quot;:[{&quot;nodeName&quot;:&quot;COMPONENT&quot;\,&quot;value&quot;:{&quot;id&quot;:&quot;$form.button1667450213112&quot;\,&quot;type&quot;:&quot;BUTTON&quot;\,&quot;name&quot;:&quot;button1667450213112&quot;}}\,{&quot;nodeName&quot;:&quot;EVENT_AND_COMPARISON_OPERATOR&quot;\,&quot;choice&quot;:{&quot;nodeName&quot;:&quot;is clicked&quot;\,&quot;value&quot;:null}}\,{&quot;nodeName&quot;:&quot;PRIMITIVE_EXPRESSION&quot;\,&quot;choice&quot;:null}]}\,&quot;nested&quot;:false}\,{&quot;nodeName&quot;:&quot;Then&quot;\,&quot;value&quot;:null}\,{&quot;nodeName&quot;:&quot;BLOCK_STATEMENTS&quot;\,&quot;items&quot;:[{&quot;nodeName&quot;:&quot;BLOCK_STATEMENT&quot;\,&quot;choice&quot;:{&quot;nodeName&quot;:&quot;SUBMIT_FORM&quot;\,&quot;items&quot;:[]}}]}]}}]\,&quot;isValid&quot;:true\,&quot;enabled&quot;:true\,&quot;version&quot;:1\,&quot;script&quot;:[&quot;submitForm()&quot;]\,&quot;eventName&quot;:&quot;Click&quot;\,&quot;ruleType&quot;:&quot;&quot;\,&quot;description&quot;:&quot;&quot;}]"
+                        jcr:primaryType="nt:unstructured"/>
+                <fd:events
+                        jcr:primaryType="nt:unstructured"
+                        click="[submitForm()]"/>
+            </submit>
+        </guideContainer>
+    </jcr:content>
+</jcr:root>
