+++ ui.af.apps/src/main/content/jcr_root/apps/core/fd/components/form/fileinput/v1/fileinput/clientlibs/site/js/fileinputwidget.js
@@ -19,7 +19,7 @@
  * file list, handling invalid file size, file name, file mime type functionality
  */
 if (typeof window.FileInputWidget === 'undefined') {
-    window.FileInputWidget = class extends FormView.FormFileInputWidget {
+    window.FileInputWidget = class extends FormView.FormFileInputWidgetBase {
         constructor(params) {
             super(params);
         }
