+++ ui.tests/test-module/specs/fileinput/fileinputv3.runtime.spec.js
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright 2022 Adobe
+ * Copyright 2024 Adobe
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,6 +13,31 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ******************************************************************************/
+const checkFileNamesInFileAttachmentView = (component, fileNames) => {
+    // check if file present in view
+    cy.get(component).then(() => {
+        fileNames.forEach((fileName) => {
+            cy.get(".cmp-adaptiveform-fileinput__filename").contains(fileName)
+        })
+    });
+}
+
+const checkFilePreviewInFileAttachment = (component) => {
+    cy.get(component).then(() => {
+        cy.get(".cmp-adaptiveform-fileinput__filename").each(($file) => {
+            cy.wrap($file).click();
+            cy.window().its('open').should('be.called');
+        })
+    });
+};
+
+const deleteSelectedFiles = (component, fileNames) => {
+    cy.get(component).then(() => {
+        fileNames.forEach((fileName) => {
+            cy.get(".cmp-adaptiveform-fileinput__filename").contains(fileName).next().find('.cmp-adaptiveform-fileinput__filedelete').click();
+        })
+    });
+};
 
 const getFormObjTest = (fileList) => {
     cy.window().then(function(win) {
@@ -39,66 +64,9 @@ const getFormObjTest = (fileList) => {
     });
 };
 
-const checkFilePreviewInFileAttachment = (component) => {
-    cy.get(component).then(() => {
-        cy.get(".cmp-adaptiveform-fileinput__filename").each(($file) => {
-            cy.wrap($file).click();
-            cy.window().its('open').should('be.called');
-        })
-    });
-};
+describe("Form with File Input V-3 - Basic Tests", () => {
 
-const deleteSelectedFiles = (component, fileNames) => {
-    cy.get(component).then(() => {
-        fileNames.forEach((fileName) => {
-            cy.get(".cmp-adaptiveform-fileinput__filedelete").should('have.attr', 'role', 'button');
-            cy.get(".cmp-adaptiveform-fileinput__filename").contains(fileName).next().click();
-        })
-    });
-};
-
-const checkFileNamesInFileAttachmentView = (component, fileNames) => {
-    // check if file present in view
-    cy.get(component).then(() => {
-        fileNames.forEach((fileName) => {
-            cy.get(".cmp-adaptiveform-fileinput__filename").contains(fileName)
-        })
-    });
-}
-
-const multiSelectFileInput1 = "input[name='fileinput1']";
-const singleSelectFileInput2 = "input[name='fileinput2']";
-const singleSelectFileInput4 = "input[name='fileinput4']";
-
-const fileInputs = [
-    {
-        type: "Single Select",
-        selector: singleSelectFileInput2,
-        fileNames: ['empty.pdf'], multiple: false
-    },
-    {
-        type: "Multi Select",
-        selector: multiSelectFileInput1,
-        fileNames: ['empty.pdf', 'sample.txt'], multiple: true
-    }
-];
-
-const unknownMimetypesFileInputs = [
-    {
-        type: "Single Select",
-        selector: singleSelectFileInput4,
-        fileNames: ['test.bat'], multiple: false
-
-    },
-    {
-        type: "Single Select",
-        selector: singleSelectFileInput4,
-        fileNames: ['test.msg'], multiple: false
-    }]    
-
-describe("Form with File Input - Basic Tests", () => {
-
-    const pagePath = "/content/forms/af/core-components-it/samples/fileinput/fileinputv1/basic.html"
+    const pagePath = "/content/forms/af/core-components-it/samples/fileinput/fileinputv3/basic.html"
     const bemBlock = 'cmp-adaptiveform-fileinput'
     const IS = "adaptiveFormFileInput"
     const selectors = {
@@ -225,20 +193,6 @@ describe("Form with File Input - Basic Tests", () => {
         })
     })
 
-    it("check preview and delete functionality of duplicate files", () => {
-        let sampleFileNames = ['sample2.txt','sample.txt','sample2.txt'];
-        const fileInput = "input[name='fileinput1']";
-        cy.get(fileInput).attachFile(sampleFileNames[0]);
-        cy.get(fileInput).attachFile(sampleFileNames[1]);
-        cy.get(fileInput).attachFile(sampleFileNames[0]);
-
-        checkFilePreviewInFileAttachment(fileInput);
-
-        deleteSelectedFiles(fileInput, sampleFileNames)
-
-        cy.get('.cmp-adaptiveform-fileinput__filelist').eq(0).children().should('have.length', 0);
-    })
-
     it(`fielinput is disabled when readonly property is true`, () => {
         const fileInput5 =  "input[name='fileinput5']";
         cy.get(fileInput5).should("have.attr", "disabled", "disabled"); 
@@ -246,132 +200,31 @@ describe("Form with File Input - Basic Tests", () => {
 
 })
 
-describe("Form with File Input - Prefill & Submit tests", () => {
-    let prefillId;
-    const submitBtn = "submit1673953138924";
-    const pagePath = "/content/forms/af/core-components-it/samples/fileinput/fileinputv1/basic.html"
-
+describe('Click on button tag (V-3)', () => {
+    const pagePath = "/content/forms/af/core-components-it/samples/fileinput/fileinputv3/basic.html"
+    const bemBlock = 'cmp-adaptiveform-fileinput'
+    const IS = "adaptiveFormFileInput"
+    const selectors = {
+        fileinput : `[data-cmp-is="${IS}"]`
+    }
+    let formContainer = null
 
     beforeEach(() => {
-        cy.wrap(prefillId).as('prefillId');
-        cy.intercept({
-            method: 'POST',
-            url: '**/adobe/forms/af/submit/*',
-        }).as('afSubmission')
+        cy.previewForm(pagePath, {
+            onBeforeLoad : (win) => {
+                cy.stub(win, 'open'); // creating a stub to check file preview
+            }
+        }).then(p => {
+            formContainer = p;
+        });
     });
-
-    const submitTest = () => {
-        cy.wait('@afSubmission').then(({response}) => {
-            const { body } = response;
-            assert.equal(response.statusCode, 200, "submission failed since status code is not 200");
-            assert.equal(body.redirectUrl.indexOf("guideThankYouPage") !== -1, true, "url does not contain thank you page in submission response");
-            assert.isDefined(body?.thankYouMessage)
-            assert.isDefined(body?.metadata?.prefillId, "prefillId not present")
-            prefillId = body.metadata.prefillId;
-            cy.wrap(prefillId).as("prefillId");
-        })
-    };
-
-    unknownMimetypesFileInputs.forEach((fileInput, idx) => {
-        it(`${fileInput.type} - attach files, check model, view, preview attachment and submit the form with bat and msg extension files`, () => {
-            cy.previewForm(pagePath, {
-                onBeforeLoad: (win) => {
-                    cy.stub(win, 'open'); // creating a stub to check file preview
-                }
-            });
-
-            // attach the file
-            cy.get(fileInput.selector).attachFile(fileInput.fileNames);
-
-            // check for successful attachment of file in the view
-            checkFileNamesInFileAttachmentView(fileInput.selector, fileInput.fileNames);
-
-            // check preview of the file
-            checkFilePreviewInFileAttachment(fileInput.selector);
-        })
-    })
-
-    fileInputs.forEach((fileInput, idx) => {
-        it(`${fileInput.type} - attach files, check model, view, preview attachment and submit the form`, () => {
-            cy.previewForm(pagePath, {
-                onBeforeLoad: (win) => {
-                    cy.stub(win, 'open'); // creating a stub to check file preview
-                }
-            });
-
-            // attach the file
-            cy.get(fileInput.selector).attachFile(fileInput.fileNames);
-            if(fileInput.multiple)
-                cy.get(fileInput.selector).attachFile('sample2.txt');
-
-            // check for successful attachment of file in the view
-            checkFileNamesInFileAttachmentView(fileInput.selector, fileInput.fileNames);
-            if(fileInput.multiple)
-                checkFileNamesInFileAttachmentView(fileInput.selector, ['sample2.txt']);
-
-            // check preview of the file
-            checkFilePreviewInFileAttachment(fileInput.selector);
-
-            if(fileInput.multiple)
-                deleteSelectedFiles(fileInput.selector, ['sample2.txt']);
-
-            // submit the form
-            cy.get(".cmp-adaptiveform-button__widget").click();
-        
-            // check for successful submission
-            submitTest();
-        })
-
-        it(`${fileInput.type} - view prefill of submitted form, make changes to attachments and submit`, () => {
-            cy.get("@prefillId").then(id => {
-                cy.previewForm(pagePath, {
-                    params: [`prefillId=${id}`],
-                    onBeforeLoad(win) {
-                        cy.stub(win, 'open'); // creating a stub to check file preview
-                    }
-                });
-
-                // check if files were prefilled
-                checkFileNamesInFileAttachmentView(fileInput.selector, fileInput.fileNames);
-
-                // check if guideBridge API returns file attachments correctly
-                getFormObjTest(['empty.pdf', ...(fileInput.multiple ? ['sample.txt']: []) ]);
-
-                // check the preview of the file attachment
-                checkFilePreviewInFileAttachment(fileInput.selector);
-
-                // add new files after preview to both the component
-                cy.get(fileInput.selector).attachFile(['sample2.txt'])
-
-                // check if guideBridge API returns correctly after prefill and attaching more files
-                getFormObjTest(['sample2.txt', ...(fileInput.multiple ? ['sample.txt', 'empty.pdf']: []) ]);
-
-                // submit the form
-                cy.get(".cmp-adaptiveform-button__widget").click();
-
-                // check if submission is success
-                submitTest();
-            });
+    it('should handle click event on attach button', () => {
+        const sampleFileName = 'sample2.txt'
+        const [id, fieldView] = Object.entries(formContainer._fields)[2]
+        cy.get(`#${id} > .${bemBlock}__container > .${bemBlock}__dragarea > .${bemBlock}__widgetlabel`).should('exist').click().then(() => {
+            cy.get('input[type="file"]').attachFile(sampleFileName).then(() => {
+                cy.get('.cmp-adaptiveform-fileinput__filename').should('contain', sampleFileName);
         });
-
-        it(`${fileInput.type} - prefill of submitted prefilled form`, () => {
-            cy.get("@prefillId").then(id => {
-                cy.previewForm(pagePath, {
-                    params: [`prefillId=${id}`],
-                    onBeforeLoad(win) {
-                        cy.stub(win, 'open'); // creating a stub to check file preview
-                    }
-                });
-
-                // check if files were prefilled
-                checkFileNamesInFileAttachmentView(fileInput.selector, ['sample2.txt', ...(fileInput.multiple ? ['sample.txt', 'empty.pdf']: []) ]);
-                getFormObjTest(['sample2.txt', ...(fileInput.multiple ? ['sample.txt', 'empty.pdf']: []) ]);
-
-                // check if file preview works fine after prefill
-                checkFilePreviewInFileAttachment(fileInput.selector);
-            });
-
-        });
-    })
-
-});
+    });
+  }); 
+})
